+++ Assets/Scripts/NPC Scripts/EnemyNPC.cs
@@ -15,10 +15,10 @@ public class EnemyNPC : MonoBehaviour {
 
     [HideInInspector] public CharacterState currentState;
 
-    [HideInInspector] public int currentStrength;
-    [HideInInspector] public int currentIntelligence;
-    [HideInInspector] public int currentResistance;
-    [HideInInspector] public int currentDexterity;
+    [HideInInspector] public float currentStrength;
+    [HideInInspector] public float currentIntelligence;
+    [HideInInspector] public float currentResistance;
+    [HideInInspector] public float currentDexterity;
 
     private int timePassed = 0;
     [HideInInspector] public int pacifier = 1;
@@ -109,7 +109,7 @@ public class EnemyNPC : MonoBehaviour {
 
         combatController.UpdateEnemyLog("El " + myTemplate.npcName + " ha atacado.");
 
-        int damage = currentStrength + Random.Range(1, 5) + Random.Range(0, 3);
+        float damage = currentStrength + Random.Range(1, 5) + Random.Range(0, 3);
 
         float r = Random.Range(0f, 1f);
